FIND-251:added,updated,relocated tests
