+++ test/lib/test_forms_fields.py
@@ -1,7 +1,6 @@
 from app.lib.fields import (
     CharField,
     ChoiceField,
-    DynamicMultipleChoiceField,
 )
 from app.lib.forms import BaseForm
 from django.http import QueryDict
@@ -240,156 +239,6 @@ class BaseFormWithChoiceFieldTest(TestCase):
                     "[yes ] is not one of the available choices. "
                     "Valid choices are [yes, no]"
                 )
-            },
-        )
-
-
-class BaseFormWithDynamicMultipleChoiceFieldTest(TestCase):
-
-    def get_form_with_dynamic_multiple_choice_field(self, data=None):
-
-        class MyTestForm(BaseForm):
-            def add_fields(self):
-                return {
-                    "dmc_field": DynamicMultipleChoiceField(
-                        label="Location",
-                        choices=[
-                            ("london", "London (100 recs)"),
-                            ("leeds", "Leeds (50 recs)"),
-                        ],
-                        required=True,
-                        validate_input=True,
-                    )
-                }
-
-        form = MyTestForm(data)
-        return form
-
-    def test_form_with_dymanic_multiple_choice_field_initial_attrs(self):
-
-        form = self.get_form_with_dynamic_multiple_choice_field()
-        self.assertEqual(form.fields["dmc_field"].name, "dmc_field")
-        self.assertEqual(form.fields["dmc_field"].label, "Location")
-        self.assertEqual(form.fields["dmc_field"].hint, "")
-
-    def test_form_with_dynamic_multiple_choice_field_with_no_params(self):
-
-        data = QueryDict("")
-        form = self.get_form_with_dynamic_multiple_choice_field(data)
-        valid_status = form.is_valid()
-        self.assertEqual(valid_status, False)
-        self.assertEqual(
-            form.errors, {"dmc_field": {"text": "Value is required."}}
-        )
-        self.assertEqual(form.fields["dmc_field"].value, [])
-        self.assertEqual(form.fields["dmc_field"].cleaned, None)
-        self.assertEqual(
-            form.fields["dmc_field"].items,
-            [
-                {"text": "London (100 recs)", "value": "london"},
-                {"text": "Leeds (50 recs)", "value": "leeds"},
-            ],
-        )
-        self.assertEqual(
-            form.fields["dmc_field"].error, {"text": "Value is required."}
-        )
-
-    def test_form_with_dynamic_multiple_choice_field_with_param_with_valid_value(
-        self,
-    ):
-
-        data = QueryDict("dmc_field=london")
-        form = self.get_form_with_dynamic_multiple_choice_field(data)
-        valid_status = form.is_valid()
-        self.assertEqual(valid_status, True)
-        self.assertEqual(form.errors, {})
-        self.assertEqual(form.fields["dmc_field"].value, ["london"])
-        self.assertEqual(form.fields["dmc_field"].cleaned, ["london"])
-        self.assertEqual(
-            form.fields["dmc_field"].items,
-            [
-                {
-                    "text": "London (100 recs)",
-                    "value": "london",
-                    "checked": True,
-                },
-                {"text": "Leeds (50 recs)", "value": "leeds"},
-            ],
-        )
-        self.assertEqual(form.fields["dmc_field"].error, {})
-
-    def test_form_with_dynamic_multiple_choice_field_with_multiple_param_with_valid_values(
-        self,
-    ):
-
-        data = QueryDict("dmc_field=london&dmc_field=leeds")
-        form = self.get_form_with_dynamic_multiple_choice_field(data)
-        valid_status = form.is_valid()
-        self.assertEqual(valid_status, True)
-        self.assertEqual(form.errors, {})
-        self.assertEqual(form.fields["dmc_field"].name, "dmc_field")
-        self.assertEqual(form.fields["dmc_field"].value, ["london", "leeds"])
-        self.assertEqual(form.fields["dmc_field"].cleaned, ["london", "leeds"])
-        self.assertEqual(
-            form.fields["dmc_field"].items,
-            [
-                {
-                    "text": "London (100 recs)",
-                    "value": "london",
-                    "checked": True,
-                },
-                {
-                    "text": "Leeds (50 recs)",
-                    "value": "leeds",
-                    "checked": True,
-                },
-            ],
-        )
-        self.assertEqual(form.fields["dmc_field"].error, {})
-
-    def test_form_with_dynamic_multiple_choice_field_with_multiple_param_with_invalid_values(
-        self,
-    ):
-
-        data = QueryDict("dmc_field=london&dmc_field=manchester")
-        form = self.get_form_with_dynamic_multiple_choice_field(data)
-        valid_status = form.is_valid()
-        self.assertEqual(valid_status, False)
-        self.assertEqual(
-            form.errors,
-            {
-                "dmc_field": {
-                    "text": (
-                        "Enter a valid choice. Value(s) [london, manchester] do not belong "
-                        "to the available choices. Valid choices are [london, leeds]"
-                    )
-                }
-            },
-        )
-        self.assertEqual(
-            form.fields["dmc_field"].value, ["london", "manchester"]
-        )
-        self.assertEqual(
-            form.fields["dmc_field"].items,
-            [
-                {
-                    "text": "London (100 recs)",
-                    "value": "london",
-                    "checked": True,
-                },
-                {
-                    "text": "Leeds (50 recs)",
-                    "value": "leeds",
-                },
-            ],
-        )
-        self.assertEqual(
-            form.fields["dmc_field"].error,
-            {
-                "text": (
-                    "Enter a valid choice. Value(s) [london, manchester] do not belong "
-                    "to the available choices. Valid choices are [london, leeds]"
-                ),
             },
         )
